chore: rename teritori-dapp -> teritori and gno-dapp -> gnotribe
