+++ apps/teritori/app.config.js
